moves tracer to a separate package
