+++ network/p2p/tracer/gossipSubMeshTracer.go
@@ -1,4 +1,4 @@
-package p2pnode
+package tracer
 
 import (
 	"sync"
